deploy changes

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -34,7 +34,7 @@ int interpolation_search(int *array, size_t size, int value)
 		key = low + (((double)(high - low) / (array[high] - array[low])) * (value -
 					array[low]));
 		printf("Value checked array[%lu] = [%d]\n",
-				(unsigned long)key, array[key]);
+				key, array[key]);
 
 		if (array[key] == value)
 		{
@@ -50,6 +50,6 @@ int interpolation_search(int *array, size_t size, int value)
 		}
 	} while (low <= high && value >= array[low] && value <= array[high]);
 
-	printf("Value checked array[%lu] is out of range\n", (unsigned long)high);
+	printf("Value checked array[%lu] is out of range\n", high);
 	return (-1);
 }
